Add data class model Restaurant Add empty usecase List RestaurantUseCase Add empty ListRestaurantsViewModel Add tests for future GooglePlacesMappingRepository in GooglePlacesMappingRepositoryTest

+++ app/src/main/java/com/gt/go4lunch/data/Location.kt
@@ -7,9 +7,9 @@ class Location {
 
     @SerializedName("lat")
     @Expose
-    var lat: Double? = null
+    var lat: Float? = null
     @SerializedName("lng")
     @Expose
-    var lng: Double? = null
+    var lng: Float? = null
 
 }
